refactor: merge helper files together in one helper index page and helper folder
