+++ src/components/TeacherComponents/SearchResultsItem/SearchResultsItem.stories.tsx
@@ -6,7 +6,7 @@ import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
 import elasticResponseFixture from "@/context/Search/search-api/2023/elasticResponse.2023.fixture.json";
 import keyStagesFixture from "@/node-lib/curriculum-api-2023/fixtures/keyStages.fixture";
 import { searchResultsHitSchema } from "@/context/Search/search.schema";
-import { getSearchHitObject } from "@/context/Search/search.helpers";
+import { getSearchHitObject } from "@/app/(core)/teachers/search/helpers/index";
 
 export default {
   decorators: [AnalyticsDecorator],
