+++ src/components/TeacherComponents/SearchResultsItem/SearchResultsItem.test.tsx
@@ -4,7 +4,7 @@ import SearchResultsItem from "./SearchResultsItem";
 
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import searchPageFixture from "@/node-lib/curriculum-api-2023/fixtures/searchPage.fixture";
-import { getSearchHitObject } from "@/context/Search/search.helpers";
+import { getSearchHitObject } from "@/app/(core)/teachers/search/helpers/index";
 import { LEGACY_COHORT } from "@/config/cohort";
 import { hitsFixture } from "@/context/Search/search-api/2023/searchResults.fixture";
